feat: extract initial version from my dotfiles
Add a script that executes `nvimrc` files in the project's directory.
This allows changing Neovim configuration based on the project without
polluting the main dotfiles.

Changes to project nvimrc files are detected and the user is prompted to
choose what to do when the file changed. This is done to prevent
running arbitrary code in nvimrcs automatically.