refactor: use vim.secure module to manage trust
This simplifies the plugin by a lot. It no longer needs to manage
trusted files, since it can rely on `vim.secure.read` to manage trust
and prompt the user to approve or deny the file.

The plugin also gained a new `reset()` method that resets the trust for
the found nvimrcs and asks the user again.